extend matching in main for location

+++ src/limit.rs
@@ -8,10 +8,10 @@ use crate::geo::distance_points;
 pub enum Limit {
     /// strategy to check limit based on the number of points
     Points(u32),
-    /// strategy to check limit based on the length of the sum of the distances between the points
+    /// strategy to check limit based on the length in meter of the sum of the distances between the points
     Length(u32),
-    /// strategy to check limit based on the distance to the nearest location
-    Location(Box<Vec<Waypoint>>, f64),
+    /// strategy to check limit based on the distance in meter to the nearest location
+    Location(Box<Vec<Waypoint>>, u32),
 }
 
 impl Limit {
@@ -25,7 +25,7 @@ impl Limit {
         Limit::Length(max_length)
     }
 
-    pub fn location(waypoint_file: String, distance: f64) -> Self {
+    pub fn location(waypoint_file: String, distance: u32) -> Self {
         debug!("reading waypoints for splitting at location from: {}", waypoint_file);
         debug!("minimum distance for location to split: {}", distance);
         let waypoints: Vec<Waypoint> = vec![];
@@ -47,11 +47,11 @@ mod tests {
 
     #[test]
     fn test_location() {
-        let lim = Limit::location("waypoint_file".to_string(), 10.0);
+        let lim = Limit::location("waypoint_file".to_string(), 10);
         match lim {
             Limit::Location(waypoints, dist) => {
                 assert_eq!(0, waypoints.len());
-                assert_eq!(10.0, dist)
+                assert_eq!(10, dist)
             },
             _ => panic!("unexpected result")
         }
